Implement the Virtio driver to operate with the Linux OS
In ThreadX, the Virtio implementation originates from rpmsg-lite [1].
I reimplemented it at the environment and platform layers, managing interrupt
and synchronization functions as well as memory management.

Prior to that, in the main function, I rewrote the mailbox ISR and
set the base address of shared memory to retrieve the available and
used ring buffers. Additionally, I fixed the issue at the Virtio handshake stage
on my C906L core, where the CPU did not write through to the shared memory,
so I flushed the Dcache to resolve this problem.

[1] RPMsg-lite (https://github.com/nxp-mcuxpresso/rpmsg-lite.git)